design: 사진 등록 페이지
